examples/langchain-python/pyproject.toml
[project]
name = "langchain-python"
version = "0.1.0"
description = "LangChain Python example with code interpreter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agbcloud-sdk",
    "python-dotenv",
    "langchain",
    "langchain-openai",
]
